Moved to hooks into client api
